Move ProducerBatch to separate file
